Update install.sh to test for bookworm firmware location
Raspberry Pi OS Bookworm (12) updated the location of the firmware configuration from /boot/config.txt to /boot/firmware/config.txt. The install script needs to detect whether the new file exists, and if so, change the path of the file to modify. I didn't test this, but it looks right.